requirements/base.txt
# Everything needed in production

# Flask
click>=7.0
Flask==2.3.2
Werkzeug==2.3.6
Flask-Restful==0.3.10
flask-restx==1.1.0
Flask-HTTPAuth==4.8.0

# Database
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.6
SQLAlchemy==2.0.18

# Migrations
Flask-Migrate==4.0.4

# Forms
email-validator==2.0.0.post2
Flask-WTF==1.1.1
WTForms==3.0.1

# Deployment
gevent==22.10.2
gunicorn>=19.9.0
supervisor==4.2.5

# Flask Static Digest
Flask-Static-Digest==0.4.0

# Auth
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2

# Caching
Flask-Caching>=2.0.2

# Debug toolbar
Flask-DebugToolbar==0.13.1

# Environment variable parsing
environs==9.5.0

# Credit card validation
git+https://github.com/maistodos/python-creditcard.git@main

# Pydantic
pydantic==2.0.2

# Crypto
pycryptodome==3.18.0